+++ src/dmx/hal/gpio.c
@@ -1,6 +1,5 @@
 #include "gpio.h"
 
-#include "dmx/config.h"
 #include "dmx/hal/uart.h"
 #include "dmx/struct.h"
 #include "hal/gpio_hal.h"
