+++ src/dmx/hal/uart.c
@@ -1,6 +1,6 @@
 #include "uart.h"
 
-#include "dmx/config.h"
+#include "dmx/io.h"
 #include "dmx/struct.h"
 #include "driver/uart.h"
 #include "esp_dmx.h"
