+++ src/dmx/io.h
@@ -1,0 +1,72 @@
+/**
+ * @file driver.h
+ * @author your name (you@domain.com)
+ * @brief 
+ * @version 0.1
+ * @date 2023-10-26
+ * 
+ * @copyright Copyright (c) 2023
+ * 
+ */
+#pragma once
+
+#include "dmx_types.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/**
+ * @brief Receives a DMX packet from the DMX bus. This is a blocking function.
+ * This function first blocks until the DMX driver is idle and then it blocks
+ * using a timeout until a new packet is received. This function will timeout
+ * early according to RDM specification if an RDM packet is expected.
+ *
+ * @note This function uses FreeRTOS direct-to-task notifications to block and
+ * unblock. Using task notifications on the same task that calls this function
+ * can lead to undesired behavior and program instability.
+ *
+ * @param dmx_num The DMX port number.
+ * @param[out] packet An optional pointer to a dmx_packet_t which contains
+ * information about the received DMX packet.
+ * @param wait_ticks The number of ticks to wait before this function times out.
+ * @return The size of the received DMX packet or 0 if no packet was received.
+ */
+size_t dmx_receive(dmx_port_t dmx_num, dmx_packet_t *packet,
+                   TickType_t wait_ticks);
+
+/**
+ * @brief Sends a DMX packet on the DMX bus. This function blocks indefinitely
+ * until the DMX driver is idle and then sends a packet.
+ *
+ * @note This function uses FreeRTOS direct-to-task notifications to block and
+ * unblock. Using task notifications on the same task that calls this function
+ * can lead to undesired behavior and program instability.
+ *
+ * @param dmx_num The DMX port number.
+ * @param size The size of the packet to send. If 0, sends the number of bytes
+ * equal to the highest slot number that was written or sent in the previous
+ * call to dmx_write(), dmx_write_offset(), dmx_write_slot(), or dmx_send().
+ * @return The number of bytes sent on the DMX bus.
+ */
+size_t dmx_send(dmx_port_t dmx_num, size_t size);
+
+/**
+ * @brief Waits until the DMX packet is done being sent. This function can be
+ * used to ensure that calls to dmx_write() happen synchronously with the
+ * current DMX frame.
+ *
+ * @note This function uses FreeRTOS direct-to-task notifications to block and
+ * unblock. Using task notifications on the same task that calls this function
+ * can lead to undesired behavior and program instability.
+ *
+ * @param dmx_num The DMX port number.
+ * @param wait_ticks The number of ticks to wait before this function times out.
+ * @retval true if the DMX driver is done sending.
+ * @retval false if the function timed out.
+ */
+bool dmx_wait_sent(dmx_port_t dmx_num, TickType_t wait_ticks);
+
+#ifdef __cplusplus
+}
+#endif
