+++ src/dmx/rw.c
@@ -1,0 +1,287 @@
+#include "dmx/rw.h"
+
+#include "dmx/driver.h"
+#include "dmx/struct.h"
+#include "endian.h"
+
+size_t dmx_read_offset(dmx_port_t dmx_num, size_t offset, void *destination,
+                       size_t size) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+  DMX_CHECK(offset < DMX_PACKET_SIZE_MAX, 0, "offset error");
+  DMX_CHECK(destination, 0, "destination is null");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
+
+  // Clamp size to the maximum DMX packet size
+  if (size + offset > DMX_PACKET_SIZE_MAX) {
+    size = DMX_PACKET_SIZE_MAX - offset;
+  } else if (size == 0) {
+    return 0;
+  }
+
+  dmx_driver_t *const driver = dmx_driver[dmx_num];
+
+  // Copy data from the driver buffer to the destination asynchronously
+  memcpy(destination, driver->data + offset, size);
+
+  return size;
+}
+
+size_t dmx_read(dmx_port_t dmx_num, void *destination, size_t size) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+  DMX_CHECK(destination, 0, "destination is null");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
+
+  return dmx_read_offset(dmx_num, 0, destination, size);
+}
+
+int dmx_read_slot(dmx_port_t dmx_num, size_t slot_num) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, -1, "dmx_num error");
+  DMX_CHECK(slot_num < DMX_PACKET_SIZE_MAX, -1, "slot_num error");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), -1, "driver is not installed");
+
+  uint8_t slot;
+  dmx_read_offset(dmx_num, slot_num, &slot, 1);
+
+  return slot;
+}
+
+size_t DMX_ISR_ATTR dmx_read_rdm(dmx_port_t dmx_num, rdm_header_t *header,
+                                 void *pd, size_t num) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
+
+  size_t read = 0;
+
+  dmx_driver_t *const driver = dmx_driver[dmx_num];
+
+  // Get pointers to driver data buffer locations and declare checksum
+  uint16_t checksum = 0;
+  const uint8_t *header_ptr = driver->data;
+  const void *pd_ptr = header_ptr + 24;
+
+  // Verify start code and sub-start code are correct
+  if (*(uint16_t *)header_ptr != (RDM_SC | (RDM_SUB_SC << 8)) &&
+      *header_ptr != RDM_PREAMBLE && *header_ptr != RDM_DELIMITER) {
+    return read;
+  }
+
+  // Get and verify the preamble_len if packet is a discovery response
+  size_t preamble_len = 0;
+  if (*header_ptr == RDM_PREAMBLE || *header_ptr == RDM_DELIMITER) {
+    for (; preamble_len <= 7; ++preamble_len) {
+      if (header_ptr[preamble_len] == RDM_DELIMITER) break;
+    }
+    if (preamble_len > 7) {
+      return read;
+    }
+  }
+
+  // Handle packets differently if a DISC_UNIQUE_BRANCH packet was received
+  if (*header_ptr == RDM_SC) {
+    // Verify checksum is correct
+    const uint8_t message_len = header_ptr[2];
+    for (int i = 0; i < message_len; ++i) {
+      checksum += header_ptr[i];
+    }
+    if (checksum != bswap16(*(uint16_t *)(header_ptr + message_len))) {
+      return read;
+    }
+
+    // Copy the header and pd from the driver
+    if (header != NULL) {
+      // Copy header without emplace so this function can be used in IRAM ISR
+      for (int i = 0; i < sizeof(rdm_header_t); ++i) {
+        ((uint8_t *)header)[i] = header_ptr[i];
+      }
+      header->dest_uid.man_id = bswap16(header->dest_uid.man_id);
+      header->dest_uid.dev_id = bswap32(header->dest_uid.dev_id);
+      header->src_uid.man_id = bswap16(header->src_uid.man_id);
+      header->src_uid.dev_id = bswap32(header->src_uid.dev_id);
+      header->sub_device = bswap16(header->sub_device);
+      header->pid = bswap16(header->pid);
+    }
+    if (pd != NULL) {
+      const uint8_t pdl = header_ptr[23];
+      const size_t copy_size = pdl < num ? pdl : num;
+      memcpy(pd, pd_ptr, copy_size);
+    }
+
+    // Update the read size
+    read = message_len + 2;
+
+  } else {
+    // Verify the checksum is correct
+    header_ptr += preamble_len + 1;
+    for (int i = 0; i < 12; ++i) {
+      checksum += header_ptr[i];
+    }
+    if (checksum != (((header_ptr[12] & header_ptr[13]) << 8) |
+                     (header_ptr[14] & header_ptr[15]))) {
+      return read;
+    }
+
+    // Decode the EUID
+    uint8_t buf[6];
+    for (int i = 0, j = 0; i < 6; ++i, j += 2) {
+      buf[i] = header_ptr[j] & header_ptr[j + 1];
+    }
+
+    // Copy the data into the header
+    if (header != NULL) {
+      // Copy header without emplace so this function can be used in IRAM ISR
+      for (int i = 0; i < sizeof(rdm_uid_t); ++i) {
+        ((uint8_t *)&header->src_uid)[i] = buf[i];
+      }
+      header->src_uid.man_id = bswap16(header->src_uid.man_id);
+      header->src_uid.dev_id = bswap32(header->src_uid.dev_id);
+      header->dest_uid = (rdm_uid_t){0, 0};
+      header->tn = 0;
+      header->response_type = RDM_RESPONSE_TYPE_ACK;
+      header->message_count = 0;
+      header->sub_device = RDM_SUB_DEVICE_ROOT;
+      header->cc = RDM_CC_DISC_COMMAND_RESPONSE;
+      header->pid = RDM_PID_DISC_UNIQUE_BRANCH;
+      header->pdl = preamble_len + 1 + 16;
+    }
+
+    // Update the read size
+    read = preamble_len + 1 + 16;
+  }
+
+  return read;
+}
+
+size_t dmx_write_offset(dmx_port_t dmx_num, size_t offset, const void *source,
+                        size_t size) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+  DMX_CHECK(offset < DMX_PACKET_SIZE_MAX, 0, "offset error");
+  DMX_CHECK(source, 0, "source is null");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
+
+  // Clamp size to the maximum DMX packet size
+  if (size + offset > DMX_PACKET_SIZE_MAX) {
+    size = DMX_PACKET_SIZE_MAX - offset;
+  } else if (size == 0) {
+    return 0;
+  }
+
+  dmx_driver_t *const driver = dmx_driver[dmx_num];
+
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  if ((driver->flags & DMX_FLAGS_DRIVER_IS_SENDING) && driver->rdm_type != 0) {
+    // Do not allow asynchronous writes when sending an RDM packet
+    taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+    return 0;
+  } else if (dmx_uart_get_rts(driver->uart) == 1) {
+    // Flip the bus to stop writes from being overwritten by incoming data
+    dmx_uart_set_rts(driver->uart, 0);
+  }
+  driver->tx_size = offset + size;  // Update driver transmit size
+
+  // Copy data from the source to the driver buffer asynchronously
+  memcpy(driver->data + offset, source, size);
+
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+
+  return size;
+}
+
+size_t dmx_write(dmx_port_t dmx_num, const void *source, size_t size) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+  DMX_CHECK(source, 0, "source is null");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
+
+  return dmx_write_offset(dmx_num, 0, source, size);
+}
+
+int dmx_write_slot(dmx_port_t dmx_num, size_t slot_num, uint8_t value) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, -1, "dmx_num error");
+  DMX_CHECK(slot_num < DMX_PACKET_SIZE_MAX, -1, "slot_num error");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), -1, "driver is not installed");
+
+  dmx_write_offset(dmx_num, slot_num, &value, 1);
+
+  return value;
+}
+
+size_t dmx_write_rdm(dmx_port_t dmx_num, rdm_header_t *header, const void *pd) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+  DMX_CHECK(header != NULL || pd != NULL, 0,
+            "header is null and pd does not contain a UID");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
+
+  size_t written = 0;
+
+  dmx_driver_t *const driver = dmx_driver[dmx_num];
+
+  // Get pointers to driver data buffer locations and declare checksum
+  uint16_t checksum = 0;
+  uint8_t *header_ptr = driver->data;
+  void *pd_ptr = header_ptr + 24;
+
+  // RDM writes must be synchronous to prevent data corruption
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  if (driver->flags & DMX_FLAGS_DRIVER_IS_SENDING) {
+    taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+    return written;
+  } else if (dmx_uart_get_rts(driver->uart) == 1) {
+    dmx_uart_set_rts(driver->uart, 0);  // Stops writes from being overwritten
+  }
+
+  if (header != NULL && !(header->cc == RDM_CC_DISC_COMMAND_RESPONSE &&
+                          header->pid == RDM_PID_DISC_UNIQUE_BRANCH)) {
+    // Copy the header, pd, message_len, and pdl into the driver
+    const size_t copy_size = header->pdl <= 231 ? header->pdl : 231;
+    header->message_len = copy_size + 24;
+    rdm_pd_emplace(header_ptr, "#cc01hbuubbbwbwb", header, sizeof(*header),
+                   false);
+    memcpy(pd_ptr, pd, copy_size);
+
+    // Calculate and copy the checksum
+    checksum = RDM_SC + RDM_SUB_SC;
+    for (int i = 2; i < header->message_len; ++i) {
+      checksum += header_ptr[i];
+    }
+    *(uint16_t *)(header_ptr + header->message_len) = bswap16(checksum);
+
+    // Update written size
+    written = header->message_len + 2;
+  } else {
+    // Encode the preamble bytes
+    const size_t preamble_len = 7;
+    for (int i = 0; i < preamble_len; ++i) {
+      header_ptr[i] = RDM_PREAMBLE;
+    }
+    header_ptr[preamble_len] = RDM_DELIMITER;
+    header_ptr += preamble_len + 1;
+
+    // Encode the UID and calculate the checksum
+    uint8_t uid[6];
+    if (header == NULL) {
+      memcpy(uid, pd, sizeof(rdm_uid_t));
+    } else {
+      rdm_uidcpy(uid, &header->src_uid);
+    }
+    for (int i = 0, j = 0; j < sizeof(rdm_uid_t); i += 2, ++j) {
+      header_ptr[i] = uid[j] | 0xaa;
+      header_ptr[i + 1] = uid[j] | 0x55;
+      checksum += uid[j] + (0xaa | 0x55);
+    }
+    header_ptr += sizeof(rdm_uid_t) * 2;
+
+    // Encode the checksum
+    header_ptr[0] = (uint8_t)(checksum >> 8) | 0xaa;
+    header_ptr[1] = (uint8_t)(checksum >> 8) | 0x55;
+    header_ptr[2] = (uint8_t)(checksum) | 0xaa;
+    header_ptr[3] = (uint8_t)(checksum) | 0x55;
+
+    // Update written size
+    written = preamble_len + 1 + 16;
+  }
+
+  // Update driver transmission size
+  driver->tx_size = written;
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+
+  return written;
+}
