+++ src/dmx/rw.h
@@ -1,0 +1,124 @@
+/**
+ * @file driver.h
+ * @author your name (you@domain.com)
+ * @brief 
+ * @version 0.1
+ * @date 2023-10-26
+ * 
+ * @copyright Copyright (c) 2023
+ * 
+ */
+#pragma once
+
+#include "dmx_types.h"
+#include "rdm_types.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/**
+ * @brief Reads DMX data from the driver into a destination buffer with an
+ * offset. This can be useful when a receiving DMX device only needs to process
+ * a small footprint of the DMX packet.
+ *
+ * @param dmx_num The DMX port number.
+ * @param offset The number of slots with which to offset the read. If set to 0
+ * this function is equivalent to dmx_read().
+ * @param[out] destination The destination buffer into which to read the DMX
+ * data.
+ * @param size The size of the destination buffer.
+ * @return The number of bytes read from the DMX driver.
+ */
+size_t dmx_read_offset(dmx_port_t dmx_num, size_t offset, void *destination,
+                       size_t size);
+
+/**
+ * @brief Reads DMX data from the driver into a destination buffer.
+ *
+ * @param dmx_num The DMX port number.
+ * @param[out] destination The destination buffer into which to read the DMX
+ * data.
+ * @param size The size of the destination buffer.
+ * @return The number of bytes read from the DMX driver.
+ */
+size_t dmx_read(dmx_port_t dmx_num, void *destination, size_t size);
+
+/**
+ * @brief Reads a single slot of DMX data.
+ *
+ * @param dmx_num The DMX port number.
+ * @param slot_num The DMX slot number to read.
+ * @return The value of the DMX slot or -1 on error.
+ */
+int dmx_read_slot(dmx_port_t dmx_num, size_t slot_num);
+
+/**
+ * @brief Reads an RDM packet from the DMX driver buffer. Header information is
+ * emplaced into a header pointer so that it may be read by the caller.
+ * Parameter data information needs to be emplaced before it can be properly
+ * read by the caller. This function does not perform any data error checking to
+ * ensure that the RDM packet is within specification.
+ *
+ * @param dmx_num The DMX port number.
+ * @param[out] header A pointer which stores RDM header information.
+ * @param[out] pd A pointer to store parameter data from the RDM packet.
+ * @param num The size of the pd pointer. Used to prevent buffer overflows.
+ * @return The size of the RDM packet that was read or 0 on error.
+ */
+size_t dmx_read_rdm(dmx_port_t dmx_num, rdm_header_t *header, void *pd,
+                    size_t num);
+
+/**
+ * @brief Writes DMX data from a source buffer into the DMX driver buffer with
+ * an offset. Allows a source buffer to be written to a specific slot number in
+ * the DMX driver buffer.
+ *
+ * @param dmx_num The DMX port number.
+ * @param offset The number of slots with which to offset the write. If set to 0
+ * this function is equivalent to dmx_write().
+ * @param[in] source The source buffer which is copied to the DMX driver.
+ * @param size The size of the source buffer.
+ * @return The number of bytes written into the DMX driver.
+ */
+size_t dmx_write_offset(dmx_port_t dmx_num, size_t offset, const void *source,
+                        size_t size);
+
+/**
+ * @brief Writes DMX data from a source buffer into the DMX driver buffer. Data
+ * written into the DMX driver buffer can then be sent to DMX devices.
+ *
+ * @param dmx_num The DMX port number.
+ * @param[in] source The source buffer which is copied to the DMX driver.
+ * @param size The size of the source buffer.
+ * @return The number of bytes written into the DMX driver.
+ */
+size_t dmx_write(dmx_port_t dmx_num, const void *source, size_t size);
+
+/**
+ * @brief Writes a single slot of DMX data.
+ *
+ * @param dmx_num The DMX port number.
+ * @param slot_num The DMX slot number to write.
+ * @return The value written to the DMX slot or -1 on error.
+ */
+int dmx_write_slot(dmx_port_t dmx_num, size_t slot_num, uint8_t value);
+
+/**
+ * @brief Writes an RDM packet into the DMX driver buffer so it may be sent with
+ * dmx_send(). Header information is emplaced into the DMX driver buffer but
+ * parameter data information must be emplaced before calling this function to
+ * ensure that the RDM packet is properly formatted. This function does not
+ * perform any data error checking to ensure that the RDM packet is within
+ * specification.
+ *
+ * @param dmx_num The DMX port number.
+ * @param[in] header A pointer which stores RDM header information.
+ * @param[in] pd A pointer which stores parameter data to be written.
+ * @return The size of the RDM packet that was written or 0 on error.
+ */
+size_t dmx_write_rdm(dmx_port_t dmx_num, rdm_header_t *header, const void *pd);
+
+#ifdef __cplusplus
+}
+#endif
