+++ src/rdm/controller.c
@@ -1,9 +1,8 @@
 #include "controller.h"
 
-#include "dmx/config.h"
+#include "dmx/driver.h"
 #include "dmx/struct.h"
 #include "endian.h"
-#include "esp_dmx.h"
 #include "rdm_utils.h"
 
 bool rdm_send_disc_unique_branch(dmx_port_t dmx_num, rdm_header_t *header,
