+++ src/rdm_utils.c
@@ -3,12 +3,15 @@
 #include <ctype.h>
 #include <string.h>
 
-#include "dmx/config.h"
 #include "dmx/hal/nvs.h"
 #include "dmx/struct.h"
 #include "endian.h"
 #include "esp_dmx.h"
 #include "rdm_types.h"
+
+#include "dmx/driver.h"
+#include "dmx/rw.h"
+#include "dmx/io.h"
 
 void *rdm_uidcpy(void *restrict destination, const void *restrict source) {
   assert(destination != NULL);
